+++ app/cells/folio/console/translated_inputs_cell.rb
@@ -26,7 +26,7 @@ class Folio::Console::TranslatedInputsCell < Folio::ConsoleCell
   end
 
   def args_with_locale(locale)
-    common = { wrapper: :with_flag, flag: locale }
+    common = { wrapper: :with_flag, flag: locale, hint: false }
 
     if args.present? && first = args.first.presence
       [