+++ app/helpers/folio/console/index_helper.rb
@@ -4,7 +4,7 @@ module Folio::Console::IndexHelper
   def index_thumb(cover, link = nil)
     return nil if cover.blank?
 
-    src = cover.thumb(Folio::FileSerializer::ADMIN_THUMBNAIL_SIZE).url
+    src = cover.thumb(Folio::Console::FileSerializer::ADMIN_THUMBNAIL_SIZE).url
     img = image_tag(src, class: 'folio-console-index-table__img')
     if link
       link_to(img, link)