+++ config/initializers/dragonfly.rb
@@ -90,12 +90,12 @@ Dragonfly.app.configure do
               server_root: Rails.root.join('public')
   else
     datastore :s3,
-            bucket_name: ENV.fetch('S3_BUCKET_NAME'),
-            access_key_id: ENV.fetch('AWS_ACCESS_KEY_ID'),
-            secret_access_key: ENV.fetch('AWS_SECRET_ACCESS_KEY'),
-            url_scheme: ENV.fetch('S3_SCHEME'),
-            region: ENV.fetch('S3_REGION'),
-            root_path: "#{ENV.fetch('PROJECT_NAME')}/#{Rails.env}/files",
-            fog_storage_options: { path_style: true }
+              bucket_name: ENV.fetch('S3_BUCKET_NAME'),
+              access_key_id: ENV.fetch('AWS_ACCESS_KEY_ID'),
+              secret_access_key: ENV.fetch('AWS_SECRET_ACCESS_KEY'),
+              url_scheme: ENV.fetch('S3_SCHEME'),
+              region: ENV.fetch('S3_REGION'),
+              root_path: "#{ENV.fetch('PROJECT_NAME')}/#{Rails.env}/files",
+              fog_storage_options: { path_style: true }
   end
 end