+++ lib/folio/console_cell.rb
@@ -3,8 +3,14 @@
 class Folio::ConsoleCell < Folio::ApplicationCell
   include Folio::Console::CellsHelper
 
-  def html_safe_fields_for(f, key, &block)
-    f.simple_fields_for key do |subfields|
+  def html_safe_fields_for(f, key, build: false, &block)
+    obj = f.object.send(key)
+
+    if build && obj.blank?
+      obj = f.object.send("build_#{key}")
+    end
+
+    f.simple_fields_for key, obj do |subfields|
       (yield subfields).html_safe
     end
   end