+++ react/src/components/FileList/FileTableList/FileTableRow/index.js
@@ -8,11 +8,11 @@ import FileUploadProgress from 'components/FileUploadProgress';
 
 const FileTableRow = ({ file, link, fileTypeIsImage, overflowingParent, onClick }) => {
   let className = 'folio-console-file-table__tr'
-  const persistedOnClick = !file.uploading && onClick
+  const persistedOnClick = !file.attributes.uploading && onClick
 
-  if (file.freshlyUploaded) {
+  if (file.attributes.freshlyUploaded) {
     className = 'folio-console-file-table__tr folio-console-file-table__tr--fresh'
-  } else if (file.uploading) {
+  } else if (file.attributes.uploading) {
     className = 'folio-console-file-table__tr folio-console-file-table__tr--uploading'
   }
 
@@ -23,19 +23,19 @@ const FileTableRow = ({ file, link, fileTypeIsImage, overflowingParent, onClick 
     >
       {fileTypeIsImage && (
         <div className='folio-console-file-table__td folio-console-file-table__td--image'>
-          <FileUploadProgress progress={file.progress} />
+          <FileUploadProgress progress={file.attributes.progress} />
 
           <div className='folio-console-file-table__img-wrap'>
-            {file.thumb && (
+            {file.attributes.thumb && (
               <a
-                href={file.source_image}
+                href={file.attributes.source_image}
                 target='_blank'
                 className='folio-console-file-table__img-a'
                 rel='noopener noreferrer'
                 onClick={(e) => e.stopPropagation()}
               >
                 <LazyLoad height={50} once overflow={overflowingParent}>
-                  <img src={file.thumb} className='folio-console-file-table__img' alt='' />
+                  <img src={file.attributes.thumb} className='folio-console-file-table__img' alt='' />
                 </LazyLoad>
               </a>
             )}
@@ -44,22 +44,22 @@ const FileTableRow = ({ file, link, fileTypeIsImage, overflowingParent, onClick 
       )}
 
       <div className='folio-console-file-table__td folio-console-file-table__td--main'>
-        {fileTypeIsImage ? null : <FileUploadProgress progress={file.progress} />}
+        {fileTypeIsImage ? null : <FileUploadProgress progress={file.attributes.progress} />}
 
-        {link ? (
+        {(link && file.links) ? (
           <a
-            href={file.edit_path}
+            href={file.links.edit}
             onClick={(e) => e.stopPropagation()}
           >
-            {file.file_name}
+            {file.attributes.file_name}
           </a>
-        ) : file.file_name}
+        ) : file.attributes.file_name}
       </div>
       <div className='folio-console-file-table__td folio-console-file-table__td--tags'>
         <Tags file={file} />
       </div>
-      <div className='folio-console-file-table__td folio-console-file-table__td--size'>{numberToHumanSize(file.file_size)}</div>
-      <div className='folio-console-file-table__td folio-console-file-table__td--extension'>{file.extension}</div>
+      <div className='folio-console-file-table__td folio-console-file-table__td--size'>{numberToHumanSize(file.attributes.file_size)}</div>
+      <div className='folio-console-file-table__td folio-console-file-table__td--extension'>{file.attributes.extension}</div>
     </div>
   )
 }