+++ react/src/containers/ModalSelect/ModalSingleSelect/index.js
@@ -22,8 +22,8 @@ class ModalSingleSelect extends ModalSelect {
       return `
         <div class="folio-console-thumbnail__inner">
           <i class="folio-console-thumbnail__fa-icon fa fa-file-o"></i>
-          <strong class="folio-console-thumbnail__title">${truncate(file.file_name)}</strong>
-          <input type="hidden" name="${prefix}[title]" value="" data-file-name="${file.file_name}" />
+          <strong class="folio-console-thumbnail__title">${truncate(file.attributes.file_name)}</strong>
+          <input type="hidden" name="${prefix}[title]" value="" data-file-name="${file.attributes.file_name}" />
           <div class="folio-console-hover-destroy">
             <i class="fa fa-edit folio-console-thumbnail__title-edit"></i>
             <i class="fa fa-times-circle" data-destroy-association></i>
@@ -34,7 +34,7 @@ class ModalSingleSelect extends ModalSelect {
       return `
         <div class="folio-console-thumbnail__inner">
           <div class="folio-console-thumbnail__img-wrap">
-            <img class="folio-console-thumbnail__img" src=${window.encodeURI(file.thumb)} alt="" />
+            <img class="folio-console-thumbnail__img" src=${window.encodeURI(file.attributes.thumb)} alt="" />
             <div class="folio-console-hover-destroy">
               <i class="fa fa-times-circle" data-destroy-association></i>
             </div>